Moved faker-map tests into superstring.test.js and updated benchmark test result location
